Added bin/dawn-git-build
This may be moved into a controller instead.
This file will build the repo using buildstep (or another build system, when and if its changed)
Its meant to operate alone without the need for the server,
as it's called using Process.spawn (truth be told, I haven't tried it yet).

tl;dr
Builds images, subject to change.